Build 2016.10.10
* end of rewrite clsFrame -> clsFrame2. Major performance gains.
* tested with
** regular animal
** dolphin's underwater shadows
** Restaurant (idle view had an issue, weird empty sort of frame + extra
bg frame; used was fine. No idea why it worked in previous versions of
ZT Studio without this  new implementation). Tried recoloring.
** cinema (one of the largest / most colored objects in the game. Or so
it appears. Anything bigger? This one lead to a few memory issues when
hovering/detecting color. Fixed!)
** showbuoy (show area in MM, largest width in the game?)
** testbox (Blue Fang image)
** bamboo (plain and simple)
** bluepath
** small concrete fence

* color palettes trigger some re-renderings
** moving colors = regenerate hex
** if a palette is replaced, we regenerate the image. Alternative would
have been to regenerate the hex, but it doesn't make sense for recolors

* (batch) offsetting now works

Loading PNGs: shouldn't this be better? Right now we still give the full
image rather than the cropped version. Might need to further adjust
offsets for it.

After batch PNG to ZT1 of Red Panda, and after the rotation fix (16
pixels up), the animal seems to be fine ingame. But the icon when you
select the animal, is too much to the bottom. Related to cropping?